LIN-2415: Fix title and tick color on step 1
Also extract most of the paywall view into its own subview to have a more module view.
Do the same for the title.
